requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.6
django-bootstrap4==2.0.1
django-crispy==1.8.1
django-crispy-forms==1.9.1
django-proxy==1.2.1
idna==2.9
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
imutils==0.5.3
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
numpy==1.18.4
opencv-contrib-python==4.2.0.34
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pyzbar==0.1.8
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
